rewrite, major mustache compat, html, more
added conditional blocks, as opposed to sections, which evaluate js expression instead of property truthyness. added html escaper with new {{=path}} injector. changed object iteration to use a fixed KEY instead of a custom key (for compat with this. adder). added a "this." prefixer to most normal injections to simply using the pure function-based internal core.